refactor: :zap: change TypeWriter to Smart and Dumb component concept and improve performance
