+++ components/TypeWriter/TypeWriter.tsx
@@ -1,0 +1,17 @@
+import styles from './styles.module.css'
+
+interface TypeWriterProps {
+  prefix?: string
+  text: string
+}
+
+const TypeWriter = ({ prefix, text }: TypeWriterProps) => {
+  return (
+    <span className={styles.typewriter}>
+      {prefix}
+      <b className="ml-2.5">{`${text}`}</b>
+    </span>
+  )
+}
+
+export default TypeWriter
